(PC-13327)[ADAGE] refactor: move filters tags test in another file
